requirements.txt
awsebcli==3.5.4
blessed==1.9.5
botocore==1.2.5
cement==2.4.0
colorama==0.3.3
configparser==3.3.0.post2
Django==1.8.4
django-bootstrap3==6.2.2
django-filter==0.11.0
django-pipeline==1.5.4
djangorestframework==3.2.4
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
docutils==0.12
futures==3.0.3
jmespath==0.8.0
Markdown==2.6.2
pathspec==0.3.3
psycopg2==2.6.1
PyExecJS==1.1.0
PyReact==0.5.2
python-dateutil==2.4.2
PyYAML==3.11
requests==2.6.2
six==1.9.0
texttable==0.8.3
wcwidth==0.1.5
websocket-client==0.32.0
wheel==0.24.0
